Add playlists related classes
