+++ Pocket Casts TV App/UI/Playlists/PlaylistDetailsViewModel.swift
@@ -3,7 +3,7 @@ import Combine
 import PocketCastsUtils
 
 @Observable
-class PlaylistDetailViewModel {
+class PlaylistDetailsViewModel {
 
     private var cancellable: AnyCancellable?
