+++ podcasts/HapticsHelper.swift
@@ -29,6 +29,20 @@ class HapticsHelper {
         triggerImpactOccurredHaptic(style: .heavy)
     }
 
+    #if os(tvOS)
+    enum FeedbackStyle {
+        case heavy
+        case light
+        case medium
+    }
+    private class func triggerImpactOccurredHaptic(style: HapticsHelper.FeedbackStyle) {
+        //No op
+    }
+
+    private class func triggerSuccessHaptic() {
+        //No op
+    }
+    #else
     private class func triggerImpactOccurredHaptic(style: UIImpactFeedbackGenerator.FeedbackStyle) {
         let feedbackGenerator = UIImpactFeedbackGenerator(style: style)
         feedbackGenerator.impactOccurred()
@@ -38,4 +52,5 @@ class HapticsHelper {
         let feedbackGenerator = UINotificationFeedbackGenerator()
         feedbackGenerator.notificationOccurred(.success)
     }
+    #endif
 }
