+++ podcasts/New Detail/Playlist Header/PlaylistHeaderViewCell.swift
@@ -1,21 +1,6 @@
 import UIKit
 import SwiftUI
 
-class PlaylistHeaderViewCellPlaceholder: ListItem {
-    override var differenceIdentifier: String {
-        "playlistHeaderResult"
-    }
-
-    static func == (lhs: PlaylistHeaderViewCellPlaceholder, rhs: PlaylistHeaderViewCellPlaceholder) -> Bool {
-        lhs.handleIsEqual(rhs)
-    }
-
-    override func handleIsEqual(_ otherItem: ListItem) -> Bool {
-        otherItem is PlaylistHeaderViewCellPlaceholder
-    }
-}
-
-
 class PlaylistHeaderViewCell: ThemeableCell {
     static let reuseIdentifier = "PlaylistHeaderViewCellIdentifier"
