+++ podcasts/New Detail/PlaylistDetailViewModel/PlaylistDetailViewModel+Archive.swift
@@ -1,3 +1,5 @@
+import PocketCastsDataModel
+
 extension PlaylistDetailViewModel {
     var shouldShowArchived: Bool {
         playlist.showArchivedEpisodes
