+++ podcasts/New Detail/PlaylistDetailViewModel/PlaylistDetailViewModel.swift
@@ -1,4 +1,5 @@
 import Foundation
+import Combine
 import PocketCastsDataModel
 import PocketCastsUtils
 import DifferenceKit
